solution: solve all compile-mode exercises

+++ exercises/primitive_types/primitive_types5/main.go
@@ -3,19 +3,18 @@
 //
 // Use Go's real numeric type names.
 
-// I AM NOT DONE
 package main
 
 import "fmt"
 
 func main() {
-	var n1 integer = 101
+	var n1 int = 101
 	if n1 > 100 {
 		fmt.Println("It is a big number")
 	} else {
 		fmt.Println("Not a big number at all")
 	}
 
-	var n2 float = 0.99
+	var n2 float64 = 0.99
 	fmt.Println(n2)
 }
